requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
tensorflow==2.14.0
numpy==1.24.3
scikit-learn==1.3.2
pyngrok==7.0.4
python-dotenv==1.0.0
requests==2.31.0